Account for mode-line and minibuffer
`which-key-posframe--max-dimensions' calculates the dimensions based
on `frame-height' and `frame-width' alone.  Doing it properly would
mean taking the value of `which-key-posframe-poshandler' into account.
Different poshandlers have different constrains.

This commit continues using the one-size-fits-all approach, but tweaks
it to account for one known defect.

`posframe-poshandler-frame-bottom-right-corner' places the posframe
above the mode-line and minibuffer of the parent frame, as do a few
others; but `which-key-posframe--max-dimensions' only subtracts one
from `frame-height'.  Change it to subtract two instead.