Shorten long lines and use spaces for indentation
In the process also make `which-key-posframe-mode' more readable by
setting each variable using its own `setq'.  This makes it easier to
see what is a variable that is being set and what is a value.  That
in turn makes it possible to put variables and values on separate
lines, which not only shortens lines but also makes it more apparent
what is going on.

Because these changes touch all the lines would be affected by a
switch from using tabs for indentation to using spaces, do that in
this commit too.  Nowadays spaces are generally preferred in Elisp
files.  Only one additional line has to be touched to do this, which
would not be touched by the other changes anyway.